refactor(api): extract stop bulk writer
Move batched stop sync writes into a dedicated StopBulkWriterService so StopPersistenceService can focus on sync flow, relation lookup, progress reporting, and deactivation.

Register the new writer in SyncModule for Nest injection.
