+++ apps/api/src/sync/stop-persistence.service.ts
@@ -1,19 +1,11 @@
 import { Injectable } from '@nestjs/common'
 import type { CityNameType } from '@bus/shared'
-import { randomUUID } from 'node:crypto'
-import {
-  DB_BEARING_BY_PRISMA,
-  DB_CITY_NAME_BY_PRISMA,
-  DB_ROUTE_SHAPE_SOURCE_BY_PRISMA,
-} from '../constants/enum-mappings.js'
-import { Prisma } from '../generated/prisma/client.js'
 import { PrismaService } from '../prisma/prisma.service.js'
-import {
-  BearingType as PrismaBearingType,
-  CityNameType as PrismaCityNameType,
-  RouteShapeSource as PrismaRouteShapeSource,
-} from '../generated/prisma/enums.js'
 import type { StopSyncRecords } from './mappers/stop.mapper.js'
+import {
+  StopBulkWriterService,
+  type AddressRecord,
+} from './stop-bulk-writer.service.js'
 import { createProgressCounter } from './sync-progress.js'
 import type { SyncResult } from './sync-result.js'
 
@@ -34,16 +26,12 @@ interface PersistStopsOptions {
   ) => Promise<void>
 }
 
-const BULK_WRITE_BATCH_SIZE = 500
-
-interface AddressRecord {
-  address_zh_tw: string | null
-  address_en: string | null
-}
-
 @Injectable()
 export class StopPersistenceService {
-  constructor(private readonly prismaService: PrismaService) {}
+  constructor(
+    private readonly prismaService: PrismaService,
+    private readonly stopBulkWriterService: StopBulkWriterService,
+  ) {}
 
   async persistStops(
     records: StopSyncRecords,
@@ -89,69 +77,19 @@ export class StopPersistenceService {
 
     options.onStageStart?.('station_groups', stationGroups.length)
 
-    for (const batch of this.chunk(stationGroups)) {
-      await this.prismaService.$executeRaw`
-        INSERT INTO "station_group" (
-          "id",
-          "uuid",
-          "tdx_station_group_id",
-          "city",
-          "name_zh_tw",
-          "name_en",
-          "name_ja",
-          "name_ko",
-          "latitude",
-          "longitude",
-          "tdx_updated_at",
-          "is_active",
-          "inactive_at",
-          "updated_at"
-        )
-        VALUES ${Prisma.join(
-          batch.map(
-            (record) => Prisma.sql`(
-              ${randomUUID()}::uuid,
-              ${record.uuid},
-              ${record.tdx_station_group_id},
-              ${this.dbCity(record.city)}::"CityNameType",
-              ${record.name_zh_tw},
-              ${record.name_en},
-              ${record.name_ja},
-              ${record.name_ko},
-              ${record.latitude},
-              ${record.longitude},
-              ${record.tdx_updated_at},
-              true,
-              NULL,
-              now()
-            )`,
-          ),
-        )}
-        ON CONFLICT ("uuid") DO UPDATE SET
-          "tdx_station_group_id" = EXCLUDED."tdx_station_group_id",
-          "city" = EXCLUDED."city",
-          "name_zh_tw" = EXCLUDED."name_zh_tw",
-          "name_en" = EXCLUDED."name_en",
-          "name_ja" = EXCLUDED."name_ja",
-          "name_ko" = EXCLUDED."name_ko",
-          "latitude" = EXCLUDED."latitude",
-          "longitude" = EXCLUDED."longitude",
-          "tdx_updated_at" = EXCLUDED."tdx_updated_at",
-          "is_active" = true,
-          "inactive_at" = NULL,
-          "updated_at" = now()
-      `
-
-      persistedCount += batch.length
-
-      if (options.onProgress && progress.shouldReport(persistedCount)) {
-        await options.onProgress(
+    await this.stopBulkWriterService.upsertStationGroups(
+      stationGroups,
+      async (batchCount) => {
+        persistedCount += batchCount
+        await this.reportProgress(
+          options,
           'station_groups',
+          progress,
           persistedCount,
           stationGroups.length,
         )
-      }
-    }
+      },
+    )
 
     const inactiveAt = new Date()
     await this.prismaService.stationGroup.updateMany({
@@ -180,85 +118,21 @@ export class StopPersistenceService {
 
     options.onStageStart?.('stations', stations.length)
 
-    for (const batch of this.chunk(stations)) {
-      await this.prismaService.$executeRaw`
-        INSERT INTO "station" (
-          "id",
-          "uuid",
-          "tdx_station_id",
-          "station_group_id",
-          "city",
-          "name_zh_tw",
-          "name_en",
-          "name_ja",
-          "name_ko",
-          "address_zh_tw",
-          "address_en",
-          "latitude",
-          "longitude",
-          "bearing",
-          "tdx_updated_at",
-          "is_active",
-          "inactive_at",
-          "updated_at"
+    await this.stopBulkWriterService.upsertStations(
+      stations,
+      stationGroupIds,
+      existingAddresses,
+      async (batchCount) => {
+        persistedCount += batchCount
+        await this.reportProgress(
+          options,
+          'stations',
+          progress,
+          persistedCount,
+          stations.length,
         )
-        VALUES ${Prisma.join(
-          batch.map((record) => {
-            const stationGroupId = record.station_group_uuid
-              ? (stationGroupIds.get(record.station_group_uuid) ?? null)
-              : null
-            const address = this.resolveAddress(
-              existingAddresses.get(record.uuid),
-              record.address_zh_tw,
-            )
-
-            return Prisma.sql`(
-              ${randomUUID()}::uuid,
-              ${record.uuid},
-              ${record.tdx_station_id},
-              ${stationGroupId}::uuid,
-              ${this.dbCity(record.city)}::"CityNameType",
-              ${record.name_zh_tw},
-              ${record.name_en},
-              ${record.name_ja},
-              ${record.name_ko},
-              ${address.address_zh_tw},
-              ${address.address_en},
-              ${record.latitude},
-              ${record.longitude},
-              ${this.dbBearing(record.bearing)}::"BearingType",
-              ${record.tdx_updated_at},
-              true,
-              NULL,
-              now()
-            )`
-          }),
-        )}
-        ON CONFLICT ("uuid") DO UPDATE SET
-          "tdx_station_id" = EXCLUDED."tdx_station_id",
-          "station_group_id" = EXCLUDED."station_group_id",
-          "city" = EXCLUDED."city",
-          "name_zh_tw" = EXCLUDED."name_zh_tw",
-          "name_en" = EXCLUDED."name_en",
-          "name_ja" = EXCLUDED."name_ja",
-          "name_ko" = EXCLUDED."name_ko",
-          "address_zh_tw" = EXCLUDED."address_zh_tw",
-          "address_en" = EXCLUDED."address_en",
-          "latitude" = EXCLUDED."latitude",
-          "longitude" = EXCLUDED."longitude",
-          "bearing" = EXCLUDED."bearing",
-          "tdx_updated_at" = EXCLUDED."tdx_updated_at",
-          "is_active" = true,
-          "inactive_at" = NULL,
-          "updated_at" = now()
-      `
-
-      persistedCount += batch.length
-
-      if (options.onProgress && progress.shouldReport(persistedCount)) {
-        await options.onProgress('stations', persistedCount, stations.length)
-      }
-    }
+      },
+    )
 
     const inactiveAt = new Date()
     await this.prismaService.station.updateMany({
@@ -307,85 +181,21 @@ export class StopPersistenceService {
 
     let persistedCount = 0
 
-    for (const batch of this.chunk(stops)) {
-      await this.prismaService.$executeRaw`
-        INSERT INTO "stop" (
-          "id",
-          "uuid",
-          "tdx_stop_id",
-          "station_id",
-          "city",
-          "name_zh_tw",
-          "name_en",
-          "name_ja",
-          "name_ko",
-          "address_zh_tw",
-          "address_en",
-          "latitude",
-          "longitude",
-          "bearing",
-          "tdx_updated_at",
-          "is_active",
-          "inactive_at",
-          "updated_at"
+    await this.stopBulkWriterService.upsertStops(
+      stops,
+      stationIds,
+      existingAddresses,
+      async (batchCount) => {
+        persistedCount += batchCount
+        await this.reportProgress(
+          { onProgress },
+          'stops',
+          progress,
+          persistedCount,
+          stops.length,
         )
-        VALUES ${Prisma.join(
-          batch.map((record) => {
-            const stationId = record.station_tdx_id
-              ? (stationIds.get(record.station_tdx_id) ?? null)
-              : null
-            const address = this.resolveAddress(
-              existingAddresses.get(record.uuid),
-              record.address_zh_tw,
-            )
-
-            return Prisma.sql`(
-              ${randomUUID()}::uuid,
-              ${record.uuid},
-              ${record.tdx_stop_id},
-              ${stationId}::uuid,
-              ${this.dbCity(record.city)}::"CityNameType",
-              ${record.name_zh_tw},
-              ${record.name_en},
-              ${record.name_ja},
-              ${record.name_ko},
-              ${address.address_zh_tw},
-              ${address.address_en},
-              ${record.latitude},
-              ${record.longitude},
-              ${this.dbBearing(record.bearing)}::"BearingType",
-              ${record.tdx_updated_at},
-              true,
-              NULL,
-              now()
-            )`
-          }),
-        )}
-        ON CONFLICT ("uuid") DO UPDATE SET
-          "tdx_stop_id" = EXCLUDED."tdx_stop_id",
-          "station_id" = EXCLUDED."station_id",
-          "city" = EXCLUDED."city",
-          "name_zh_tw" = EXCLUDED."name_zh_tw",
-          "name_en" = EXCLUDED."name_en",
-          "name_ja" = EXCLUDED."name_ja",
-          "name_ko" = EXCLUDED."name_ko",
-          "address_zh_tw" = EXCLUDED."address_zh_tw",
-          "address_en" = EXCLUDED."address_en",
-          "latitude" = EXCLUDED."latitude",
-          "longitude" = EXCLUDED."longitude",
-          "bearing" = EXCLUDED."bearing",
-          "tdx_updated_at" = EXCLUDED."tdx_updated_at",
-          "is_active" = true,
-          "inactive_at" = NULL,
-          "updated_at" = now()
-      `
-
-      persistedCount += batch.length
-
-      if (onProgress && progress.shouldReport(persistedCount)) {
-        await onProgress('stops', persistedCount, stops.length)
-      }
-    }
+      },
+    )
 
     const inactiveAt = new Date()
     const deactivatedStops = await this.prismaService.stop.updateMany({
@@ -423,70 +233,22 @@ export class StopPersistenceService {
     const progress = createProgressCounter(routeStops.length)
     let persistedCount = 0
 
-    for (const batch of this.chunk(routeStops)) {
-      const values = batch.flatMap((record) => {
-        const subrouteId = subrouteIds.get(record.subroute_uuid)
-        const stopId = stopIds.get(record.stop_uuid)
-
-        if (!subrouteId || !stopId) return []
-
-        incomingKeys.add(`${subrouteId}:${record.sequence}`)
-
-        return [
-          {
-            subrouteId,
-            stopId,
-            sequence: record.sequence,
-            tdxUpdatedAt: record.tdx_updated_at,
-          },
-        ]
-      })
-
-      if (values.length > 0) {
-        await this.prismaService.$executeRaw`
-          INSERT INTO "route_stop" (
-            "id",
-            "subroute_id",
-            "stop_id",
-            "sequence",
-            "tdx_updated_at",
-            "is_active",
-            "inactive_at",
-            "updated_at"
-          )
-          VALUES ${Prisma.join(
-            values.map(
-              (value) => Prisma.sql`(
-                ${randomUUID()}::uuid,
-                ${value.subrouteId}::uuid,
-                ${value.stopId}::uuid,
-                ${value.sequence},
-                ${value.tdxUpdatedAt},
-                true,
-                NULL,
-                now()
-              )`,
-            ),
-          )}
-          ON CONFLICT ("subroute_id", "sequence") DO UPDATE SET
-            "stop_id" = EXCLUDED."stop_id",
-            "tdx_updated_at" = EXCLUDED."tdx_updated_at",
-            "is_active" = true,
-            "inactive_at" = NULL,
-            "updated_at" = now()
-        `
-      }
-
-      persistedCount += batch.length
-
-      if (options.onProgress && progress.shouldReport(persistedCount)) {
-        await options.onProgress(
+    await this.stopBulkWriterService.upsertRouteStops(
+      routeStops,
+      subrouteIds,
+      stopIds,
+      incomingKeys,
+      async (batchCount) => {
+        persistedCount += batchCount
+        await this.reportProgress(
+          options,
           'route_stops',
+          progress,
           persistedCount,
           routeStops.length,
         )
-      }
-    }
+      },
+    )
 
     const subrouteUuidList = [
       ...new Set(routeStops.map((record) => record.subroute_uuid)),
@@ -542,122 +304,32 @@ export class StopPersistenceService {
     const progress = createProgressCounter(routeShapes.length)
     let persistedCount = 0
 
-    for (const batch of this.chunk(routeShapes)) {
-      const values = batch.flatMap((record) => {
-        const subrouteId = subrouteIds.get(record.subroute_uuid)
-
-        if (!subrouteId) return []
-
-        return [
-          {
-            subrouteId,
-            source: record.source,
-            path: record.path,
-            tdxUpdatedAt: record.tdx_updated_at,
-          },
-        ]
-      })
-
-      if (values.length > 0) {
-        await this.prismaService.$executeRaw`
-          INSERT INTO "route_shape" (
-            "id",
-            "subroute_id",
-            "source",
-            "path",
-            "encoded_polyline",
-            "geometry",
-            "tdx_updated_at",
-            "is_active",
-            "inactive_at",
-            "updated_at"
-          )
-          VALUES ${Prisma.join(
-            values.map(
-              (value) => Prisma.sql`(
-                ${randomUUID()}::uuid,
-                ${value.subrouteId}::uuid,
-                ${this.dbRouteShapeSource(value.source)}::"RouteShapeSource",
-                ${JSON.stringify(value.path)}::jsonb,
-                NULL,
-                NULL,
-                ${value.tdxUpdatedAt},
-                true,
-                NULL,
-                now()
-              )`,
-            ),
-          )}
-          ON CONFLICT ("subroute_id") DO UPDATE SET
-            "source" = EXCLUDED."source",
-            "path" = EXCLUDED."path",
-            "encoded_polyline" = NULL,
-            "geometry" = NULL,
-            "tdx_updated_at" = EXCLUDED."tdx_updated_at",
-            "is_active" = true,
-            "inactive_at" = NULL,
-            "updated_at" = now()
-          WHERE "route_shape"."source" = 'stop_positions'::"RouteShapeSource"
-        `
-      }
-
-      persistedCount += batch.length
-
-      if (options.onProgress && progress.shouldReport(persistedCount)) {
-        await options.onProgress(
+    await this.stopBulkWriterService.upsertRouteShapes(
+      routeShapes,
+      subrouteIds,
+      async (batchCount) => {
+        persistedCount += batchCount
+        await this.reportProgress(
+          options,
           'route_shapes',
+          progress,
           persistedCount,
           routeShapes.length,
         )
-      }
-    }
+      },
+    )
   }
 
-  private resolveAddress(
-    existing: AddressRecord | undefined,
-    incomingAddressZhTw: string | null,
-  ): AddressRecord {
-    if (
-      existing &&
-      existing.address_zh_tw === incomingAddressZhTw &&
-      existing.address_en
-    ) {
-      return {
-        address_zh_tw: incomingAddressZhTw,
-        address_en: existing.address_en,
-      }
-    }
+  private async reportProgress(
+    options: Pick<PersistStopsOptions, 'onProgress'>,
+    stage: StopSyncStage,
+    progress: ReturnType<typeof createProgressCounter>,
+    persistedCount: number,
+    totalCount: number,
+  ): Promise<void> {
+    if (!options.onProgress || !progress.shouldReport(persistedCount)) return
 
-    return {
-      address_zh_tw: incomingAddressZhTw,
-      address_en: null,
-    }
-  }
-
-  private chunk<T>(records: T[]): T[][] {
-    const chunks: T[][] = []
-
-    for (
-      let index = 0;
-      index < records.length;
-      index += BULK_WRITE_BATCH_SIZE
-    ) {
-      chunks.push(records.slice(index, index + BULK_WRITE_BATCH_SIZE))
-    }
-
-    return chunks
-  }
-
-  private dbCity(city: PrismaCityNameType): string {
-    return DB_CITY_NAME_BY_PRISMA[city]
-  }
-
-  private dbBearing(bearing: PrismaBearingType | null): string | null {
-    return bearing ? DB_BEARING_BY_PRISMA[bearing] : null
-  }
-
-  private dbRouteShapeSource(source: PrismaRouteShapeSource): string {
-    return DB_ROUTE_SHAPE_SOURCE_BY_PRISMA[source]
+    await options.onProgress(stage, persistedCount, totalCount)
   }
 
   private async loadStationGroupIds(
