+++ apps/api/src/sync/sync.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common'
 import { PrismaModule } from '../prisma/prisma.module.js'
 import { RoutePersistenceService } from './route-persistence.service.js'
 import { RoutesSyncService } from './routes-sync.service.js'
+import { StopBulkWriterService } from './stop-bulk-writer.service.js'
 import { StopPersistenceService } from './stop-persistence.service.js'
 import { StopsSyncService } from './stops-sync.service.js'
 import { SyncCheckpointService } from './sync-checkpoint.service.js'
@@ -13,6 +14,7 @@ import { TdxClientService } from './tdx-client.service.js'
   providers: [
     RoutePersistenceService,
     RoutesSyncService,
+    StopBulkWriterService,
     StopPersistenceService,
     StopsSyncService,
     SyncCheckpointService,
